requirements.txt
python-telegram-bot==20.0
langchain
langchain-community
langchain-openai
openai
langgraph
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.92.0
python-dotenv
pytest==7.4.0
pytest-asyncio==0.20.3